Added new parser
Added new parser script that will be used for the actual cs2 research, currently working on formatting it and getting it to a good state. Currently it can collect fields parse it and add it all to one csv file, however i am currently working on formatting better to our needs.